Created common component for material-ui tooltip with modified format
